requirements.txt
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
durationpy==0.5
Flask==2.0.3
Flask-WTF==1.1.1
hjson==3.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
placebo==0.9.0
cparser==0.2.0
pymongo[srv]==3.13.0
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
urllib3==1.26.16
Werkzeug==2.3.6
WTForms==3.0.1
zappa==0.57.0
